backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
reportlab==4.0.7
google-generativeai==0.3.0
python-dotenv==1.0.0
pydantic>=2.0,<3.0
flask-cors==4.0.0
